Ensure that abbreviations stick together
This is sort of an edge case: If the abbreviation doesn't backspace (so
the abbreviation stays visible), and it doesn't wait for a completion
character, but sends the last char, we have to wait for that char to
send so we can send the expansion. Otherwise we end up with the
abbreviation getting broken up by the expansion.