Change back to string keys
This reverts commit 28519278ddcfff23d39e5c7bd4990dd8bc130151.

The rationale behind this is that this allows arbitrary precision: We'll
never int-overflow a set, no matter how big it is, and since the strings
are interned, we shouldn't have too many of them.

One drawback is that the table addresses might change from under it (!).

For comparison, here are the other schemes I've tried to avoid
stringing (for n nodes):

- Each node is numbered with a sequential ID

  Nodes are combined by sorting their IDs, then joining them by
  summing id[i]*n^i

  The IDs must be sorted to make the same sets in different orders
  return the same key. The maximum value of a set's key is n^m where m is the
  number of nodes in the set, and n is the number of nodes in the trie.

- Each node is numbered with an ID that doubles for each

  This essentially makes each one a bit in a bitfield, so they can be
  joined by summing without ordering.

  However, the maximum value is 2^(n+1)-1, and even a one-node set can
  be 2^n, so for any trie with >64 nodes, we start seeing wraparound
  fast.