Fix internal triggering
The problem was that we weren't handling internal state transitions
inside the character loop, so `c` was undefined (nil) for that piece of
code.

Instead, clarify that state transitions are done per character code.
Indeed, _expansions_ are done per character code, because if any node
along the way has expansions, we want to see those, instead of maybe
skipping over them because they occurred in the middle of an event.

I'm not sure all the cases where we can get more than one character in
an event, but deadkeys are one, so I've noted that in the code.

Break out "get next state" and "process expansion" to make things a
little more modular: Code was difficult to follow and intertwined
around.