Place comparison function in expansion metatable
Instead of passing the function around, put the info in the expansion
itself. Instead of using a "less than" function, which was neat but
awkward, use a custom "supersedes" function that is a little clearer in
usage.

Also lock the defaults at start time, so changing object defaults won't
change anything until the object is reinitialized. (I don't know what it
would do, but I don't think it would be pretty.)