Don't key expansions by their abbreviations internally
It's handy to declare them that way, especially in the simple case, but
we might get a little speedup from storing them this way during build
time.

Note that the order of them is unpredictable (due to the first pairs()
iterator's lack of control over order), so don't read into that.