Fix a bug with word endings at the start
When there was an internal expansion with an end char as its first
character, ending another abbreviation with that character wouldn't
trigger anything. This is because the exact match of the character
overrode the fuzzier match of the end char.

To fix this, divorce the decision about failing to another state from
the decision about what expansions are triggered. This removes the hacky
"isCompletion" flag that persisted between edge follows, which is nice.
Instead, have the ability to have present an expansion from one completion state
while the actual state is different.