Remove function getMatchingExpansion
The parent divorced expansions from the current state. Instead of
tracking the state separately from the expansion, don't track the
expansion in the walker; just compute it once and return it at the time
of edge following.

This hinges on the fact that we will always want to know the expansion
that results from following an edge. It is, admittedly, a violation of
CQRS, for the record.